sw_scale: Add specializations for hscale 8 to 19
Add arm64 neon implementations for hscale 8 to 19 with filter
sized 4 and 8. Both implementations are based on very similar ones
dedicated to hscale 8 to 15. Tha major changes refer to saving
the data - instead of wrinting the result as int16_t it is done
with uint32_t.

The tests and benchmarks run on AWS Gravtion 3 instances. The results
are shown below.

<Fill me in please>

Signed-off-by: Hubert Mazur <hum@semihalf.com>